Rename lcd_puts_p to lcd_puts_P
In avr-libc the *_P functions accept a PGM_P (pointer to program
memory), so lcd_puts("foo") would be written lcd_puts_P(PSTR("foo")).
PSTR() puts the passed string into PROGMEM and lcd_puts_P reads it back
from there.

The LCD lib used to have a lcd_puts_p which accepts a PGM_P and a
macro, lcd_puts_P, that converts the passed string into a PSTR.

This is inconsistent with the rest of avr-libc, so change it.